refactor gc/perf functions
