+++ libs/base/control.ts
@@ -65,18 +65,6 @@ namespace control {
     export function panic(code: number) { }
 
     /**
-     * Enable profiling for current function.
-     */
-    //% shim=TD_NOOP shimArgument=perfCounter
-    export function enablePerfCounter(name?: string) { }
-
-    /**
-     * Dump values of profiling performance counters.
-     */
-    //% shim=pxt::dumpPerfCounters
-    export function dmesgPerfCounters() { }
-
-    /**
      * Display an error code and stop the program when the assertion is `false`.
      */
     //% help=control/assert weight=30
